# docs/swagger.yaml
basePath: /
definitions:
...
  models.Team:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  models.TeamMember:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      role:
        type: string
      team_id:
        type: string
    type: object
...
  models.UpdateTeam:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
  utils.AppError:
    properties:
      code:
        description: HTTP status code
        type: integer
      message:
        description: error message
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: A simple task tracker API built with Go and MongoDB
  title: Task Tracker API
  version: "1.0"
paths:
...
  /teams:
    get:
      description: Retrieve all teams for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Team'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.AppError'
      summary: Get user teams
      tags:
      - teams
    post:
      consumes:
      - application/json
      description: Create a new team
      parameters:
      - description: Team payload
        in: body
        name: team
        required: true
        schema:
          $ref: '#/definitions/models.Team'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.AppError'
      summary: Create team
      tags:
      - teams
  /teams/{id}:
    delete:
      description: Delete a team and all associated data
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.AppError'
      summary: Delete team
      tags:
      - teams
    get:
      description: Retrieve a single team by ID
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Team'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.AppError'
      summary: Get team
      tags:
      - teams
    put:
      consumes:
      - application/json
      description: Update a team by ID
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated team payload
        in: body
        name: team
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTeam'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.AppError'
      summary: Update team
      tags:
      - teams
  /teams/{id}/members:
    post:
      consumes:
      - application/json
      description: Add a new member to a team using email and role
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: string
      - description: Member payload
        in: body
        name: member
        required: true
        schema:
          $ref: '#/definitions/models.TeamMember'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.AppError'
      summary: Add member to team
      tags:
      - teams
  /teams/{id}/members/{memberID}:
    delete:
      description: Remove a member from a team
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: string
      - description: Member ID
        in: path
        name: memberID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.AppError'
      summary: Remove team member
      tags:
      - teams
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"